apps/accounts/migrations/0003_sitesettings.py
from django.db import migrations, models


class Migration(migrations.Migration):

    dependencies = [
        ('accounts', '0002_user_group'),
    ]

    operations = [
        migrations.CreateModel(
            name='SiteSettings',
            fields=[
                ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                ('allow_signup', models.BooleanField(default=False)),
            ],
            options={
                'verbose_name': 'Site Settings',
            },
        ),
    ]
