requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pytest-django>=4.8.0
coverage>=7.5.0
factory-boy>=3.3.0
